Move service/common/telemetry service/common/telemetry
